+++ lib/src/annotation/head.dart
@@ -4,6 +4,11 @@
 part of annotation;
 
 /**
+ * Short version of [Head].
+ */
+const HEAD = const Head();
+
+/**
  * Mark the annotated method to respond on HTTP HEAD requests.
  */
 class Head {