+++ lib/src/annotation/post.dart
@@ -4,6 +4,11 @@
 part of annotation;
 
 /**
+ * Short version of [Post].
+ */
+const POST = const Post();
+
+/**
  * Mark the annotated method to respond on HTTP POST requests.
  */
 class Post {